requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==24.8.1
    # via twisted
cffi==1.17.1
    # via cryptography
channels==4.2.2
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.2.1
    # via -r requirements.in
constantly==23.10.4
    # via twisted
cryptography==44.0.2
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.1.2
    # via -r requirements.in
django==5.2
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==4.7.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
drf-yasg[validation]==1.21.10
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   hyperlink
    #   twisted
importlib-resources==6.5.2
    # via swagger-spec-validator
incremental==24.7.2
    # via twisted
inflection==0.5.1
    # via drf-yasg
jsonschema==4.23.0
    # via swagger-spec-validator
jsonschema-specifications==2024.10.1
    # via jsonschema
msgpack==1.1.0
    # via channels-redis
packaging==24.2
    # via
    #   drf-yasg
    #   gunicorn
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pyopenssl==25.0.0
    # via twisted
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.2
    # via
    #   drf-yasg
    #   swagger-spec-validator
redis==5.2.1
    # via channels-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
service-identity==24.2.0
    # via twisted
sqlparse==0.5.3
    # via django
swagger-spec-validator==3.0.4
    # via drf-yasg
twisted[tls]==24.11.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.13.1
    # via
    #   pyopenssl
    #   referencing
    #   swagger-spec-validator
    #   twisted
uritemplate==4.1.1
    # via drf-yasg
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
